refactor: 불필요한 클래스 제거
- GameManager클래스를 불필요하게 한번 더 클래스로 감싸던 로직을 제거. GameManager를
제거하고 GameManager의 코드를 Game클래스로 이동.
